requirements.txt
absl-py==0.12.0
astunparse==1.6.3
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.23
fastapi==0.63.0
ffmpeg==1.4
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
h11==0.12.0
h5py==2.10.0
idna==2.10
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
onnx==1.7.0
onnxruntime==1.2.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.7.0
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
slidingwindow==0.0.14
starlette==0.13.6
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tf-slim @ git+https://github.com/adrianc-a/tf-slim.git@80265a15482b4f81f162a344f13659a855cc5543
toml==0.10.2
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
Werkzeug==1.0.1
wrapt==1.12.1
